Refactor rust_compile functions
