+++ crossbundle/tools/src/commands/android/mod.rs
@@ -17,7 +17,6 @@ mod install_apk;
 mod read_manifest;
 mod remove;
 mod rust_compile;
-mod rust_compile_macroquad;
 mod save_manifest;
 mod sign_apk;
 mod start_apk;
@@ -39,7 +38,6 @@ pub use install_apk::*;
 pub use read_manifest::*;
 pub use remove::*;
 pub use rust_compile::*;
-pub use rust_compile_macroquad::*;
 pub use save_manifest::*;
 pub use sign_apk::*;
 pub use start_apk::*;
