+++ crossbundle/tools/src/types/game_engine.rs
@@ -1,0 +1,11 @@
+#[derive(Clone, Copy)]
+pub enum GameEngine {
+    Bevy,
+    Macroquad,
+}
+
+impl Default for GameEngine {
+    fn default() -> Self {
+        Self::Macroquad
+    }
+}
