+++ crossbundle/tools/src/types/mod.rs
@@ -1,9 +1,11 @@
 mod build_target;
+mod game_engine;
 mod profile;
 mod target;
 mod version_code;
 
 pub use build_target::*;
+pub use game_engine::*;
 pub use profile::*;
 pub use target::*;
 pub use version_code::*;
